job-evaluation/pyproject.toml
[project]
name = "job-eval"
version = "0.1.0"
description = "Job classification tool for analyzing and comparing position descriptions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.9",
    "reportlab>=4.0.0",
]

[project.scripts]
job-eval = "job_eval.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]
